Python Programming Foundation
Project Description:
In the DataPrepKit capstone project, students will embark on developing a Python package named "DataPrepKit." This package aims to be a comprehensive toolkit for preprocessing datasets. Utilizing their knowledge in NumPy and Pandas, students will create a series of functions that assist in reading data from a variety of file formats, summarizing datasets, managing missing values, and encoding categorical data. The ultimate goal of this project is to publish the DataPrepKit package on PyPI, thereby making it available to the wider Python community.